requirements.txt
chromadb==1.5.0
configobj==5.0.9
configparser==7.2.0
fastapi==0.129.0
httplib2==0.31.2
httptools==0.7.1
nipype==1.10.0
ollama==0.6.1
pandas==3.0.1
pdfplumber==0.11.9
pip==24.0
pydantic-settings==2.13.1
pydentic==0.0.1.dev3
PyMuPDF==1.27.1
oython-dotenv
python-docx==1.2.0
python-multipart==0.0.22
pyxnat==1.6.4
tiktoken==0.12.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
